requirements.txt

streamlit==1.28.0
python-dotenv==1.0.0

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
langchain==0.0.350

# Embeddings and Vector Store
sentence-transformers==2.2.2
chromadb==0.4.18

# LLM Integration
ollama

# Utilities
numpy==1.24.3
pandas==2.0.3